+++ src/main/java/techcourse/myblog/service/UserService.java
@@ -4,15 +4,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import techcourse.myblog.domain.User;
 import techcourse.myblog.exception.UserDuplicateException;
-import techcourse.myblog.exception.UserMismatchException;
-import techcourse.myblog.exception.UserNotFoundException;
 import techcourse.myblog.repository.UserRepository;
 
 @Service
 public class UserService {
     private static final String ERROR_DUPLICATE_EMAIL_MESSAGE = "이미 가입된 이메일 주소입니다!";
-    private static final String ERROR_USER_NOT_FOUND_MESSAGE = "일치하는 이메일 주소가 없습니다!";
-    private static final String ERROR_MISMATCH_PASSWORD_MESSAGE = "비밀번호가 일치하지 않습니다!";
     private final UserRepository userRepository;
 
     @Autowired
@@ -35,19 +31,8 @@ public class UserService {
         return userRepository.findAll();
     }
 
-    public User findByEmail(String email) {
+    private User findByEmail(String email) {
         return userRepository.findByEmail(email);
-    }
-
-    public void checkLogin(String email, String password) {
-        if (!existsEmail(email)) {
-            throw new UserNotFoundException(ERROR_USER_NOT_FOUND_MESSAGE);
-        }
-
-        User user = findByEmail(email);
-        if (!user.matchPassword(password)) {
-            throw new UserMismatchException(ERROR_MISMATCH_PASSWORD_MESSAGE);
-        }
     }
 
     public void deleteUser(String email) {
