+++ src/main/java/techcourse/myblog/web/UserController.java
@@ -9,11 +9,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import techcourse.myblog.domain.User;
 import techcourse.myblog.dto.UserEditDto;
-import techcourse.myblog.dto.UserLoginDto;
 import techcourse.myblog.dto.UserSaveDto;
 import techcourse.myblog.exception.UserDuplicateException;
-import techcourse.myblog.exception.UserMismatchException;
-import techcourse.myblog.exception.UserNotFoundException;
 import techcourse.myblog.service.UserService;
 
 import javax.servlet.http.HttpSession;
@@ -27,40 +24,6 @@ public class UserController {
     @Autowired
     public UserController(UserService userService) {
         this.userService = userService;
-    }
-
-    @GetMapping("/login")
-    public String showLoginPage(HttpSession httpSession) {
-        if (isUserInSession(httpSession)) {
-            return "redirect:/";
-        }
-        return "login";
-    }
-
-    private boolean isUserInSession(HttpSession httpSession) {
-        return httpSession.getAttribute("user") != null;
-    }
-
-    @GetMapping("/signup")
-    public String showSignUpPage(HttpSession httpSession) {
-        if (isUserInSession(httpSession)) {
-            return "redirect:/";
-        }
-        return "signup";
-    }
-
-    @PostMapping("/login")
-    public String login(UserLoginDto userLoginDto, Model model, HttpSession httpSession) {
-        try {
-            userService.checkLogin(userLoginDto.getEmail(), userLoginDto.getPassword());
-        } catch (UserNotFoundException | UserMismatchException e) {
-            model.addAttribute("errorMessage", e.getMessage());
-            return "login";
-        }
-
-        User user = userService.findByEmail(userLoginDto.getEmail());
-        httpSession.setAttribute("user", user);
-        return "index";
     }
 
     @PostMapping("/users")
@@ -80,12 +43,6 @@ public class UserController {
 
         model.addAttribute("users", users);
         return "user-list";
-    }
-
-    @GetMapping("/logout")
-    public String logout(HttpSession httpSession) {
-        httpSession.removeAttribute("user");
-        return "redirect:/";
     }
 
     @GetMapping("/mypage")
